삭제하기 이메일 추가

+++ src/pages/api/delete/list_item.js
@@ -1,21 +1,33 @@
 import { connectDB } from "@/util/db";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 
 export default async function listItemHandler(req, res)
 {
-    console.log(req.body);
+    console.log("받은정보: ",req.body);
     if(req.method == 'DELETE')
     {
         try{
-            const {id} = req.body;
+            let session = await getServerSession(req, res, authOptions);
 
-            const db = (await connectDB).db('mydb');
-             let result = await db.collection('post').deleteOne(
-                 {_id: ObjectId.createFromHexString(id)}
-             );
-             console.log(result);
-            res.status(200).json('삭제완료');
+            const {id, email} = req.body;
+            const sessionEmail = session?.user?.email;
+            
+            if(sessionEmail === email){
+                const db = (await connectDB).db('mydb');
+                 let result = await db.collection('post').deleteOne(
+                    {
+                        _id: ObjectId.createFromHexString(id),
+                        email: sessionEmail,
+                    }
+                 );
+                 console.log(result);
+                res.status(200).json('삭제완료');
+            }else{
+                res.status(400).json({error:'계정 정보가 일치하지 않습니다'});
+            }
         }catch(error){
             console.error('Database error:', error);
             res.status(500).json({ error: 'Database Error' });
